Code cleanup and some performance improvement
Removing an extra call to Google Places service to fetch place details
that aren't needed. Also moved all the time display code to a seperate
method to simplify the getArrivals method. Also added a break to the
showArrivals for loop to exit the loop when the first matching stop is
found.